requirements.txt
streamlit
pandas
gspread
google-auth
gspread-dataframe
python-dotenv
email-validator